[front] refactor: move the comparison UI in its own component
this will allow to re-use the comparison component everywhere in
the application, for instance on the Home page, without duplicating
code
